+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on and uses the types of changes according to [Keep a Change
 ### Added
 
 - Added additional error handling to connect function to identify problematic steps when they occur.
+- Added new method of verifying spo tenant for Connect-SPOService branch of connect function.
 
 ## [0.1.26] - 2024-08-04